requirements.txt
pytest
selenium
PyHamcrest